core: Fix adb when installing a APK with incompatible abi
On AndroidO MR1, if we install a APK with abi incompatible with the
phone platform using adb of version 1.0.39, adb does not respond. Adb
commits a session and then gets result from SynchronousQueue. However
when adb gets result from SynchronousQueue ,the SynchronousQueue is
empty so adb is blocked and does not respond.

The root cause is the job of committing a session needs to write session
while a new thread holds the lock and job of calling back LocalIntentReceiver
is also on this thread.

Bug:80111622

Change-Id: Ifa0d69742a389aaaa5a3459a361c295c1984d76b
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>