Fix condition of setup always button
There will be a crash when the user clicks an item in a ResolverActivity
without AlwaysUseOption because an opposite condition was used to call
setAlwaysButtonEnabled() in ResolverActivity. TechListChooserActivity in
Nfc app is such an Activity. So the user can't choose an app when he uses
NFC with multiple apps which can handle the
NfcAdapter.ACTION_TAG_DISCOVERED.

Test: manually on Nexus 6P with a NFC card
Change-Id: Iec632fc174a693d3233512a7e04909e721b342b4