SystemUI: Port Now Playing feature from Pixel [1/2]
All of the code here was reverse engineered from smali and a
decompiled version of SystemUIGoogle from the taimen factory
images.

The decompiler wasn't a fan of the Lambdas in the code (probably
beacause they were previously obfuscated in a dex) so I converted
them into separate classes where needed.

The other commit adds the necessary files for music recognition.

I did some modification as we already have ambient music but it only shows what the device is playing and not what the device is hearing. Getting the best of both worlds

Change-Id: Ia4fd61e1101db128682efc1ef59bff8f35726028